.#3, #8, #10, #11 - 기능 합치기(관리자 기능 제외 수정필요)

+++ FoodNavi/src/main/java/com/demo/domain/FoodDetail.java
@@ -30,10 +30,10 @@ public class FoodDetail {
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	private int fdseq;
 	
-	private int kcal;
-	private int fat;
-	private int carb;
-	private int prt;
+	private float kcal;
+	private float fat;
+	private float carb;
+	private float prt;
 	
 	@OneToOne
 	@JoinColumn(name="fseq", nullable=false)
